+++ django_geoip/tests/__init__.py
@@ -18,6 +18,7 @@ unittest = my_unittest
 
 from test_management import *
 from test_base import *
+from test_storage import *
 from test_models import *
 from test_utils import *
 from test_middleware import *